+++ packages/amplify-ui-components/src/common/types.ts
@@ -1,0 +1,1 @@
+export type TextFieldTypes = 'date' | 'email' | 'number' | 'password' | 'search' | 'tel' | 'text' | 'url' | 'time';