+++ packages/amplify-ui-components/src/components.d.ts
@@ -7,6 +7,9 @@
 
 import { HTMLStencilElement, JSXBase } from '@stencil/core/internal';
 import {
+  TextFieldTypes,
+} from './common/types';
+import {
   IconNameType,
 } from './components/amplify-icon/icons';
 
@@ -23,14 +26,34 @@ export namespace Components {
   }
   interface AmplifyExamples {}
   interface AmplifyFormField {
+    /**
+    * The text of the description.  Goes between the label and the input.
+    */
+    'description': string | null;
+    /**
+    * The ID of the field.  Should match with its corresponding input's ID.
+    */
     'fieldId': string;
+    /**
+    * The text of a hint to the user as to how to fill out the input.  Goes just below the input.
+    */
     'hint': string | null;
-    'inputProps': {
-      type?: string;
-      onInput?: (Event) => void;
-    };
+    /**
+    * The text of the label.  Goes above the input. Ex: 'First name'
+    */
     'label': string | null;
-    'styleOverride': boolean;
+    /**
+    * The callback, called when the input is modified by the user.
+    */
+    'onInput'?: (inputEvent: Event) => void;
+    /**
+    * (optional) The placeholder for the input element.  Using hints is recommended, but placeholders can also be useful to convey information to users.
+    */
+    'placeholder'?: string;
+    /**
+    * The input type.  Can be any HTML input type.
+    */
+    'type'?: TextFieldTypes;
   }
   interface AmplifyHint {
     'styleOverride': boolean;
@@ -44,6 +67,28 @@ export namespace Components {
     * (Optional) Override default styling
     */
     'overrideStyle': boolean;
+  }
+  interface AmplifyInput {
+    /**
+    * The text of the description.  Goes just below the label.
+    */
+    'description': string | null;
+    /**
+    * The ID of the field.  Should match with its corresponding input's ID.
+    */
+    'fieldId': string;
+    /**
+    * The callback, called when the input is modified by the user.
+    */
+    'onInput'?: (inputEvent: Event) => void;
+    /**
+    * (optional) The placeholder for the input element.  Using hints is recommended, but placeholders can also be useful to convey information to users.
+    */
+    'placeholder'?: string;
+    /**
+    * The input type.  Can be any HTML input type.
+    */
+    'type'?: TextFieldTypes;
   }
   interface AmplifyLabel {
     'htmlFor': string;
@@ -76,7 +121,7 @@ export namespace Components {
     'fieldId': string;
     'hint': string | null;
     'inputProps': {
-      type?: string;
+      type?: TextFieldTypes;
       onChange?: (Event) => void;
     };
     'label': string | null;
@@ -87,7 +132,7 @@ export namespace Components {
     'fieldId': string;
     'hint': string | null;
     'inputProps': {
-      type?: string;
+      type?: TextFieldTypes;
       onChange?: (Event) => void;
     };
     'label': string | null;
@@ -152,6 +197,12 @@ declare global {
   var HTMLAmplifyIconElement: {
     prototype: HTMLAmplifyIconElement;
     new (): HTMLAmplifyIconElement;
+  };
+
+  interface HTMLAmplifyInputElement extends Components.AmplifyInput, HTMLStencilElement {}
+  var HTMLAmplifyInputElement: {
+    prototype: HTMLAmplifyInputElement;
+    new (): HTMLAmplifyInputElement;
   };
 
   interface HTMLAmplifyLabelElement extends Components.AmplifyLabel, HTMLStencilElement {}
@@ -226,6 +277,7 @@ declare global {
     'amplify-form-field': HTMLAmplifyFormFieldElement;
     'amplify-hint': HTMLAmplifyHintElement;
     'amplify-icon': HTMLAmplifyIconElement;
+    'amplify-input': HTMLAmplifyInputElement;
     'amplify-label': HTMLAmplifyLabelElement;
     'amplify-link': HTMLAmplifyLinkElement;
     'amplify-scene-loading': HTMLAmplifySceneLoadingElement;
@@ -254,14 +306,34 @@ declare namespace LocalJSX {
   }
   interface AmplifyExamples extends JSXBase.HTMLAttributes<HTMLAmplifyExamplesElement> {}
   interface AmplifyFormField extends JSXBase.HTMLAttributes<HTMLAmplifyFormFieldElement> {
+    /**
+    * The text of the description.  Goes between the label and the input.
+    */
+    'description'?: string | null;
+    /**
+    * The ID of the field.  Should match with its corresponding input's ID.
+    */
     'fieldId'?: string;
+    /**
+    * The text of a hint to the user as to how to fill out the input.  Goes just below the input.
+    */
     'hint'?: string | null;
-    'inputProps'?: {
-      type?: string;
-      onInput?: (Event) => void;
-    };
+    /**
+    * The text of the label.  Goes above the input. Ex: 'First name'
+    */
     'label'?: string | null;
-    'styleOverride'?: boolean;
+    /**
+    * The callback, called when the input is modified by the user.
+    */
+    'onInput'?: (inputEvent: Event) => void;
+    /**
+    * (optional) The placeholder for the input element.  Using hints is recommended, but placeholders can also be useful to convey information to users.
+    */
+    'placeholder'?: string;
+    /**
+    * The input type.  Can be any HTML input type.
+    */
+    'type'?: TextFieldTypes;
   }
   interface AmplifyHint extends JSXBase.HTMLAttributes<HTMLAmplifyHintElement> {
     'styleOverride'?: boolean;
@@ -275,6 +347,28 @@ declare namespace LocalJSX {
     * (Optional) Override default styling
     */
     'overrideStyle'?: boolean;
+  }
+  interface AmplifyInput extends JSXBase.HTMLAttributes<HTMLAmplifyInputElement> {
+    /**
+    * The text of the description.  Goes just below the label.
+    */
+    'description'?: string | null;
+    /**
+    * The ID of the field.  Should match with its corresponding input's ID.
+    */
+    'fieldId'?: string;
+    /**
+    * The callback, called when the input is modified by the user.
+    */
+    'onInput'?: (inputEvent: Event) => void;
+    /**
+    * (optional) The placeholder for the input element.  Using hints is recommended, but placeholders can also be useful to convey information to users.
+    */
+    'placeholder'?: string;
+    /**
+    * The input type.  Can be any HTML input type.
+    */
+    'type'?: TextFieldTypes;
   }
   interface AmplifyLabel extends JSXBase.HTMLAttributes<HTMLAmplifyLabelElement> {
     'htmlFor'?: string;
@@ -307,7 +401,7 @@ declare namespace LocalJSX {
     'fieldId'?: string;
     'hint'?: string | null;
     'inputProps'?: {
-      type?: string;
+      type?: TextFieldTypes;
       onChange?: (Event) => void;
     };
     'label'?: string | null;
@@ -318,7 +412,7 @@ declare namespace LocalJSX {
     'fieldId'?: string;
     'hint'?: string | null;
     'inputProps'?: {
-      type?: string;
+      type?: TextFieldTypes;
       onChange?: (Event) => void;
     };
     'label'?: string | null;
@@ -353,6 +447,7 @@ declare namespace LocalJSX {
     'amplify-form-field': AmplifyFormField;
     'amplify-hint': AmplifyHint;
     'amplify-icon': AmplifyIcon;
+    'amplify-input': AmplifyInput;
     'amplify-label': AmplifyLabel;
     'amplify-link': AmplifyLink;
     'amplify-scene-loading': AmplifySceneLoading;