+++ packages/amplify-ui-components/src/components/amplify-form-field/amplify-form-field.style.ts
@@ -1,5 +1,13 @@
 import { css } from 'emotion';
 
-export const formField = css`
-  margin-bottom: 22px;
+export const formFieldLabel = css`
+  display: block;
+  font-size: 0.9em;
+  font-weight: bold;
+  padding-bottom: 0.5em;
+`;
+
+export const formFieldDescription = css`
+  font-size: 0.8em;
+  padding-top: 0.5em;
 `;